added the requested changes: 1.moved all the views into the newly created views folder 2.fixed the api endpoints 3.fixed the normalizeData() in the api
