+++ assets/styles.css
@@ -4,6 +4,7 @@
 @import "utilities.css";
 @import "generic.css";
 @import "details.css";
+@import "home.css";
 
 * {
   box-sizing: inherit;
