+++ index.html
@@ -16,6 +16,23 @@
   </nav>
 
   <main>
+    <section id="home-section">
+      <header class="main-header">
+        <h1 class="text-center">HOME VIEW</h1>
+      </header>
+      <main class="main-info">
+        <div class="img">
+          <img src="./assets/imgs/home logo.png" alt="final-space-logo">
+        </div>
+        <div class="info">
+          <p>
+            Lorem ipsum dolor sit amet consectetur adipisicing elit. At suscipit quaerat neque repudiandae excepturi ab
+            officiis alias, vel, cumque fuga animi error minus. Est porro aperiam nemo molestiae laborum modi?
+          </p>
+          <a href="https://en.wikipedia.org/wiki/Final_Space">wikipedia</a>
+        </div>
+      </main>
+    </section>
     <!-- catalog section was here -->
     <!-- details sections was here -->
   </main>
